+++ source/common/upstream/load_stats_reporter.h
@@ -34,7 +34,7 @@ class LoadStatsReporter
 public:
   LoadStatsReporter(const envoy::api::v2::core::Node& node, ClusterManager& cluster_manager,
                     Stats::Scope& scope, Grpc::AsyncClientPtr async_client,
-                    Event::Dispatcher& dispatcher);
+                    Event::Dispatcher& dispatcher, MonotonicTimeSource& time_source);
 
   // Grpc::TypedAsyncStreamCallbacks
   void onCreateInitialMetadata(Http::HeaderMap& metadata) override;
@@ -63,7 +63,9 @@ private:
   Event::TimerPtr response_timer_;
   envoy::service::load_stats::v2::LoadStatsRequest request_;
   std::unique_ptr<envoy::service::load_stats::v2::LoadStatsResponse> message_;
-  std::vector<std::string> clusters_;
+  // Map from cluster name to start of measurement interval.
+  std::unordered_map<std::string, std::chrono::steady_clock::duration> clusters_;
+  MonotonicTimeSource& time_source_;
 };
 
 typedef std::unique_ptr<LoadStatsReporter> LoadStatsReporterPtr;