+++ source/extensions/resource_monitors/well_known_names.h
@@ -1,0 +1,23 @@
+#pragma once
+
+#include "common/singleton/const_singleton.h"
+
+namespace Envoy {
+namespace Extensions {
+namespace ResourceMonitors {
+
+/**
+ * Well-known resource monior names.
+ * NOTE: New resource monitors should use the well known name: envoy.resource_monitors.name.
+ */
+class ResourceMonitorNameValues {
+public:
+  // Heap monitor with statically configured max.
+  const std::string FIXED_HEAP_RESOURCE_MONITOR = "envoy.resource_monitors.fixed_heap";
+};
+
+typedef ConstSingleton<ResourceMonitorNameValues> ResourceMonitorNames;
+
+} // namespace ResourceMonitors
+} // namespace Extensions
+} // namespace Envoy